cloud_functions/push_to_bubble/requirements.txt
functions-framework
google-cloud-pubsub
google-cloud-logging
google-cloud-secret-manager
cloud-sql-python-connector
pg8000
sqlalchemy
jsonschema
